Refactored createTitleLine funtion. In this refactored version, the following changes have been made:
Changed the function name to createTitleLine() to adhere to camelCase naming convention.
Created separate functions initializeLineStyle() and createLineWithStyle() to simplify the code and make it more modular.
Removed unnecessary comments, as the function names and variable names should be self-explanatory.
Passed the style_line variable by reference in the initializeLineStyle() function to directly modify the style.
These changes improve the readability of the code and make it more modular, which makes it easier to maintain and modify in the future.